Reorganize files; score show when scored
